Refactor: `ptp_utils` to `prompt_to_prompt_utils`
