+++ keras_cv/models/stable_diffusion/prompt_to_prompt_utils.py
